requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv
pydantic==2.7.4
python-multipart==0.0.3
PyPDF2==3.0.1
nltk==3.9.1
transformers==4.56.0
torch==2.8.0
google.generativeai